clean up createSnapshot

+++ basebinaries/createSnapshot/main.c
@@ -1,23 +1,10 @@
-#include <unistd.h>
 #include <fcntl.h>
-#include <sys/syscall.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <strings.h>
-
-#include <unistd.h>
-#include <errno.h>
-#include <sys/attr.h>
 #include <sys/snapshot.h>
+#include <unistd.h>
 
-int
-do_create(const char *vol, const char *snap)
-{
-    int dirfd = open(vol, O_RDONLY, 0);
-    if (dirfd < 0) {
-        perror("open");
-        exit(1);
-    }
+int do_create(int dirfd, const char *snap) {
     
     int ret = fs_snapshot_create(dirfd, snap, 0);
     if (ret != 0)
@@ -25,13 +12,14 @@ do_create(const char *vol, const char *snap)
     return (ret);
 }
 
-
-int
-main(int argc, char **argv)
-{
+int main(int argc, char **argv) {
+    int dirfd = open("/", O_RDONLY, 0);
+    if (dirfd < 0) {
+        perror("open");
+        exit(1);
+    }
     unlink("/createSnapshot");
-    do_create("/", "electra-prejailbreak");
+    do_create(dirfd, "electra-prejailbreak");
     
     return (0);
 }
-
